+++ CHANGES.md
@@ -2,6 +2,7 @@ v0.2.10
 =======
 + fix extra output column in pairs.tsv (#47)
 + change output file name of ancestry to include
++ fix for gvcf with empty alts (#46)
 
 v0.2.9
 ======