+++ CHANGES.md
@@ -3,6 +3,7 @@ v0.2.10
 + fix extra output column in pairs.tsv (#47)
 + change output file name of ancestry to include
 + fix for gvcf with empty alts (#46)
++ add include regions to find-sites
 
 v0.2.9
 ======